requirements.txt
Flask==3.0.3
appengine-python-standard==1.1.6
google-cloud-ndb==2.2.1
